backend/requirements.txt
# Core
Django==5.0.7
djangorestframework==3.15.1
gunicorn==21.2.0
channels==4.0.0
daphne==4.1.0

# Environment & config
python-decouple==3.8
python-dotenv==1.0.1  # helpful for local env loading

# PostgreSQL
psycopg2-binary==2.9.9

# Authentication & Security
PyJWT==2.8.0
numpy<2.0.0

# Google Cloud
google-auth==2.29.0
google-api-python-client==2.126.0
google-cloud-translate==3.15.4

# AI Models
transformers==4.34.1
faster-whisper==0.10.0
torch==2.2.2
torchaudio==2.2.2

# Audio Processing
ffmpeg-python==0.2.0  # wrapper to control ffmpeg (optional)

# Misc
requests==2.31.0  # used in many libs
